backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.4
Pillow>=10.0
gunicorn
dj-database-url
python-dotenv
psycopg[binary]
